Move all related functions to their respective modules
