+++ src/project.js
@@ -7,14 +7,24 @@ export default class Project {
   addTodo = (todo) => {
     this.todos.push(todo);
   }
-
-  getTodo = (todoTitle) => {
-    let myTodo;
-    this.todos.forEach((todo) => {
-      if (todo.Title() === todoTitle) {
-        myTodo = todo;
-      }
-    });
-    return myTodo;
-  }
 }
+
+export const showProjects = () => {
+  const projectSection = document.getElementById('all-projects');
+  const projectsFromLocalStorage = getProjectsFromLocalStorage();
+  allProjects.push(projectsFromLocalStorage);
+  projectSection.innerHTML = '';
+  allProjects.forEach((project, idx) => {
+    const newProjectCard = projectCard(project, idx);
+    projectSection.append(newProjectCard);
+  });
+};
+
+showProjects();
+
+export const projectExists = (projectName) => {
+  allProjects = getProjectsFromLocalStorage();
+  return allProjects.find((project) => project.title === projectName);
+};
+
+
