+++ src/todo.js
@@ -6,3 +6,115 @@ export default class Todo {
     this.dueDate = dueDate;
   }
 }
+
+export const showTodos = (currentPrj, newTodoCreationForm) => {
+  todoSection.innerHTML = '';
+  allProjects = getProjectsFromLocalStorage();
+  const currentPrjIdx = allProjects.findIndex((project) => project.title === currentPrj.title);
+  currentPrj = allProjects[currentPrjIdx];
+
+  currentPrj.todos.forEach((todo) => {
+    const todoCard = document.createElement('div');
+    todoCard.className = 'todo-card';
+
+    const todoTitleProperty = document.createElement('h4');
+    todoTitleProperty.textContent = todo.title;
+    todoTitleProperty.className = 'todo-title';
+
+    const todoDescProperty = document.createElement('p');
+    todoDescProperty.textContent = todo.description;
+    todoDescProperty.className = 'todo-desc';
+
+    const todoDueDateProperty = document.createElement('p');
+    todoDueDateProperty.textContent = todo.dueDate;
+    todoDueDateProperty.className = 'todo-dueDate';
+
+    const todoPriorityProperty = document.createElement('p');
+    todoPriorityProperty.textContent = `${todo.priority}`;
+    todoPriorityProperty.className = 'todo-priority';
+
+    const modifyTodoBtn = document.createElement('button');
+    modifyTodoBtn.innerHTML = 'Edit';
+    modifyTodoBtn.className = 'update-todo-btn';
+    modifyTodoBtn.addEventListener('click', () => {
+      const newTodoForm = document.getElementById('new-todo-creation-form');
+      newTodoForm.remove();
+      modifyTodo(todo, showTodos, newTodoCreationForm);
+    });
+
+    const deleteTodoBtn = document.createElement('button');
+    deleteTodoBtn.innerText = 'delete';
+    deleteTodoBtn.id = 'todo-dlt-btn';
+    deleteTodoBtn.addEventListener('click', () => {
+      const currentTdIdx = currentPrj.todos.findIndex((eTodo) => eTodo.title === todo.title);
+      currentPrj.todos.splice(currentTdIdx, 1);
+      localStorage.setItem('allProjects', JSON.stringify(allProjects));
+      showTodos(currentPrj, newTodoCreationForm);
+    });
+
+    todoCard.append(todoTitleProperty, todoDescProperty,
+      todoPriorityProperty,
+      todoDueDateProperty, modifyTodoBtn, deleteTodoBtn);
+    todoSection.append(todoCard);
+  });
+};
+
+export const modifyTodo = (todo, showTodos, newTodoCreationForm) => {
+  const modifyTodoForm = document.createElement('form');
+  modifyTodoForm.id = 'modify-todo-form';
+  const todoTitle = document.createElement('input');
+  todoTitle.setAttribute('type', 'text');
+  todoTitle.setAttribute('placeholder', 'todo title');
+  todoTitle.id = 'todo-title-modify';
+
+  const todoDesc = document.createElement('input');
+  todoDesc.setAttribute('type', 'text');
+  todoDesc.setAttribute('placeholder', 'some description..');
+  todoDesc.id = 'todo-desc-modify';
+
+  const todoDueDate = document.createElement('input');
+  todoDueDate.setAttribute('type', 'date');
+  todoDueDate.setAttribute('placeholder', 'due date');
+  todoDueDate.id = 'todo-priority-modify';
+
+  const todoPriority = document.createElement('select');
+  todoPriority.setAttribute('id', 'priority');
+  todoPriority.id = 'todo-dueDate-modify';
+
+  const todoPriorities = ['Very high', 'High', 'Moderate', 'Low', 'Useless'];
+  const options = todoPriorities.map((priority) => {
+    const value = priority.toLowerCase();
+    return `<option value="${value}">${priority}</option>`;
+  });
+  todoPriority.innerHTML = options;
+
+  const todoSubmit = document.createElement('input');
+  todoSubmit.value = 'Update Todo';
+  todoSubmit.setAttribute('type', 'submit');
+  todoSubmit.id = 'todo-form-submit-modify';
+
+  modifyTodoForm.append(todoTitle, todoDesc, todoDueDate, todoPriority, todoSubmit);
+  projectTodos.append(modifyTodoForm);
+
+  modifyTodoForm.addEventListener('submit', (e) => {
+    e.preventDefault();
+    allProjects = getProjectsFromLocalStorage();
+    const currentPrjIdx = allProjects.findIndex(
+      (project) => project.title === currentPrj.title,
+    );
+    const currentTdIdx = allProjects[currentPrjIdx].todos.findIndex(
+      (eTodo) => eTodo.title === todo.title,
+    );
+    const currentTodo = allProjects[currentPrjIdx].todos[currentTdIdx];
+    currentTodo.title = document.getElementById('todo-title-modify').value;
+    currentTodo.description = document.getElementById('todo-desc-modify').value;
+    currentTodo.priority = document.getElementById('todo-priority-modify').value;
+    currentTodo.dueDate = document.getElementById('todo-dueDate-modify').value;
+    localStorage.setItem('allProjects', JSON.stringify(allProjects));
+    showTodos(currentPrj, newTodoCreationForm);
+    const modifyTodoForm = document.getElementById('modify-todo-form');
+    modifyTodoForm.remove();
+    newTodoCreationForm(showTodos);
+  });
+  return modifyTodoForm;
+};
